% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkdumpr-package.R
\docType{package}
\name{darkdumpr}
\alias{darkdumpr}
\alias{darkdumpr-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/darkdumpr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/darkdumpr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
