% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkdumpr.R
\name{search_noir}
\alias{search_noir}
\title{Query darkdump.io for Dark Web Content}
\usage{
search_noir(query, page = 1)
}
\arguments{
\item{query}{<chr> terms to search (supports
\href{https://darksearch.io/dorks}{darkdump.io advanced search composition}).}

\item{page}{pagination control}
}
\value{
data frame with a "\code{metadata}" atttribute containing pagination info
}
\description{
Query darkdump.io for Dark Web Content
}
\examples{
search_noir("ransomware")
}
